Header (#26)
* fix: approve commit

* fix: delete unused import

* fix: fix regis, login button. add route for profile pict. add loading

* fix: update create event, fix api for create event

* fix: fix end point for createEvent, and createTicket

* fix: update create event page

* fix: search debounce and pagination

+++ apps/api/src/controllers/profile.controller.ts
@@ -13,12 +13,21 @@ export class ProfileController {
             where: { id: res.locals.decript.id },
             data: { imageProfile: `profile/${req.file?.filename}`, }
         });
+        console.log("below this")
         console.log(join(__dirname, "../../public", `/${findUser?.imageProfile}`))
-        fs.unlinkSync(join(__dirname, "../../public", `/${findUser?.imageProfile}`)); // delete oldfile
+        // fs.unlinkSync(join(__dirname, "../../public", `/${findUser?.imageProfile}`)); // delete oldfile
+        if (fs.existsSync(join(__dirname, "../../public", `/${findUser?.imageProfile}`))) {
+            fs.unlinkSync(join(__dirname, "../../public", `/${findUser?.imageProfile}`));
+            console.log("File deleted successfully.");
+        } else {
+            console.log("File does not exist.");
+        }
         res.status(200).send({
             rc: 200,
             success: true,
             message: "Update profile success",
         })
     }
+
+
 }
